rag_agent/requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
langchain-chroma
langgraph
pypdf
chromadb
pytest
langchain-experimental
duckduckgo-search
mcp
langchain-mcp-adapters
httpx
sse-starlette
python-dotenv
exa-py
fastapi
uvicorn
